+++ src/seeds/workflow/workflow.js
@@ -1,4 +1,4 @@
-import { map } from 'lodash';
+import { map, merge } from 'lodash';
 
 import { fileExists, stripI18n } from '../../helpers';
 import { Workflow } from '../../models';
@@ -10,10 +10,21 @@ export const seedWorkflow = async (trx, profilePath) => {
       await Workflow.delete(trx);
     }
     await Promise.all(
-      map(
-        profile.workflows,
-        async (workflow) => await Workflow.create(workflow, {}, trx),
-      ),
+      map(profile.workflows, async (workflow) => {
+        // Check if type exists
+        const current = await Workflow.fetchById(workflow.id, {}, trx);
+
+        // If doesn't exist
+        if (!current) {
+          await Workflow.create(workflow, {}, trx);
+        } else {
+          await Workflow.update(
+            workflow.id,
+            merge(current.$toDatabaseJson(), workflow),
+            trx,
+          );
+        }
+      }),
     );
     console.log('Workflows imported');
   }